Remove '----' option from times drop down
Originally, I liked having a default '---------' option in the times
drop-down menu because it forced the user to look at all the available
options.

But there was a problem when a user filled out the entire form (except
for choosing a valid time). An error was raised, but the options
populated by AJAX were removed when the form was re-displayed. I
couldn't find a way to keep these prior fields without changing the
'meeting_fields' HouseholdForm field to be a ChoiceField - but that
would require introducing another Model field - which I don't want to
do.