Re-add default option to times drop down
I removed the default "-----" option from the "Select Time" dropdown in
commit d129570. This was due to a AJAX + Django form issue; if the user
did NOT select a time and submitted the form, an error was returned
(good) but the time drop down was empty upon reload (bad).

This change puts the default option back in the "Select Time" dropdown,
and ensures it's populated (even on reload) by using a jQuery document
ready AJAX call.